# src/backend/APIDocuments.yaml
openapi: 3.0.1
info:
  title: MOHLTC-DataProject
  description:  Restful APIs for MOHLTC-DataProject, mainly includes users management, data package management, and the manipulations related to the templates and the values of workbooks.
  version: '0.1'
servers:
  - url: 'http://localhost:3000'
  - url: 'http://ec2-3-13-195-83.us-east-2.compute.amazonaws.com/'
paths:
...
  /api/admin/packages/{packageName}:
    get:
      description: Returns one package
      tags:
        - Package
      parameters:
        - name: packageName
          in: path
          required: true
          schema:
            type: string
          example: 'firstpackage'
      responses:
        '404':
          description: package not found
          content:
            application/json; charset=utf-8:
              schema:
                type: string
              examples: {}
        '200':
          description: successful operation
          content:
            application/json; charset=utf-8:
              schema:
                type: string
              examples: {}
      servers:
        - url: 'http://localhost:3000'
    servers:
      - url: 'http://localhost:3000'
    put:
      description: Returen the updated the package
      tags:
        - Package
      parameters:
        - name: packageName
          in: path
          required: true
          schema:
            type: string
          example: 'firstpackage'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                organizationTypes:
                  type: array
                  items:
                    type: string
                adminNotes:
                  type: string
                workbooks:
                  type: array
                  items:
                    type: string
                endDate:
                  type: string
                organizations:
                  type: array
                  items:
                    type: string
                name:
                  type: string
                published:
                  type: boolean
                _id:
                  type: string
                startDate:
                  type: string
            examples:
              '0':
                value: |-
                  {
                              "published": false,
                              "organizations": [
                                  "5d653701adffa653832f1969"
                              ],
                              "organizationTypes": [
                                  "5d653701adffa653832f1969"
                              ],
                              "workbooks": [
                                  "5d5eb072d7ac99d4a4f473b8"
                              ],
                              "_id": "5d65380b04dbe131405f4b38",
                              "name": "firstpackage",
                              "startDate": "2019-08-27T14:02:21.360Z",
                              "endDate": "2019-09-03T14:02:21.360Z",
                              "adminNotes": "updated"
                          }
      responses:
        '200':
          description: successful operation
          content:
            application/json; charset=utf-8:
              schema:
                type: string
              examples: {}
      servers:
        - url: 'http://localhost:3000'
  /api/admin/{userName}/packages/{packageName}:
    get:
      description: Returns the packages based on the user name and package name
      tags:
        - Package
      parameters:
        - name: userName
          in: path
          required: true
          schema:
            type: string
          example: 'test'
        - name: packageName
          in: path
          required: true
          schema:
            type: string
          example: 'firstpackage'

      responses:
        '400':
          description: failed operation
          content:
            application/json; charset=utf-8:
              schema:
                type: string
              examples: {}
        '200':
          description: successful operation
          content:
            application/json; charset=utf-8:
              schema:
                type: string
              examples: {}
      servers:
        - url: 'http://localhost:3000'
    servers:
      - url: 'http://localhost:3000'
  /api/v2/admin/packages:
    get:
      description: Returns all packages on the database
      tags:
        - Package
      responses:
        '200':
          description: successful operation
          content:
            application/json; charset=utf-8:
              schema:
                type: string
              examples: {}
      servers:
        - url: 'http://localhost:3000'
    post:
      description: Create one pakcage
      tags:
        - Package
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                organizationTypes:
                  type: array
                  items:
                    type: string
                adminNotes:
                  type: string
                workbooks:
                  type: array
                  items:
                    type: string
                endDate:
                  type: string
                adminFiles:
                  type: array
                  items:
                    type: string
                organizations:
                  type: array
                  items:
                    type: string
                name:
                  type: string
                published:
                  type: boolean
                groupNumber:
                  type: integer
                startDate:
                  type: string
            examples:
              '0':
                value: |-
                  {
                              "published": true,
                              "organizations": [],
                              "organizationTypes": [],
                              "workbooks": [
                                  "5d5eb072d7ac99d4a4f473b8",
                                  "5d5eb0b8d7ac99d4a4f473f0"
                              ],
                              "name": "packageWillBeDeleted",
                              "startDate": "2019-08-22T15:12:31.541Z",
                              "endDate": "2019-08-29T15:12:31.541Z",
                              "adminNotes": "with two workbooks",
                              "adminFiles": [],
                              "groupNumber": 1
                  }
      responses:
        '200':
          description: Created successfully
          content:
            application/json; charset=utf-8:
              schema:
                type: string
              examples: {}
      servers:
        - url: 'http://localhost:3000'
    servers:
      - url: 'http://localhost:3000'
  /api/v2/admin/packages/{packageName}:
    delete:
      description: Returns one package from database
      tags:
        - Package
      parameters:
        - name: packageName
          in: path
          required: true
          schema:
            type: string
          example: 'packageWillBeDeleted'
      responses:
        '200':
          description: successful operation
          content:
            application/json; charset=utf-8:
              schema:
                type: string
              examples: {}
      servers:
        - url: 'http://localhost:3000'
    servers:
      - url: 'http://localhost:3000'
  # Workbook
...